[#205] Handle case where MAIN worker thread is asked to exit right away without waiting for a logical point to be reached (i.e. forced_simplethreadapi_true == FALSE)
A new function ydb_stm_thread_exit() is added to sr_unix/ydb_stm_thread.c.
This takes care of exit handling related activities for the MAIN worker thread.

It is invoked from ydb_stm_thread() when a logical point has been reached where it is safe to exit.
This is the case where "forced_simplethreadapi_true" is TRUE.

ydb_stm_thread_exit() is also invoked (from gtm_exit_handler()) in case "forced_simplethreadapi_exit"
is FALSE. In this case, it is possible we will never reach a logical point for safe exit (i.e. TP
worker thread can never terminate as it is waiting for MAIN worker thread to service request which
has been interrupted to handle exit handler request). To avoid deadlock in these cases, we do not
indefinitely wait, for TP worker thread to terminate, in the MAIN worker thread.  Instead we wait
for 1000 iterations of 1 microsecond each for a total of around 1 milli-second per TP worker thread
before moving on with YottaDB exit handling.

Additionally, changes to sr_port/deferred_signal_handler.c and sr_unix/generic_signal_handler.c
that deferred exit processing in case we are the MAIN worker thread (done in a prior commit
305fe6970cf6f2916e892bfba406bab4b5954eb4) are now reverted. This is because it is possible the MAIN
worker thread is servicing a long running command (e.g. ydb_ci_t of a call-in M program that runs
for ever until stopped using a kill -15 as is done in the dual_fail_extend/dual_fail2_mustop_sigquit
subtest). In that case if we defer the signal, the MAIN worker thread that is in ydb_ci_t() will
never come back to ydb_stm_threadq_dispatch() which means the process will never terminate if we
defer exit handling.  With the introduction of ydb_stm_thread_exit() in the current commit, the
MAIN worker thread will wait for TP worker threads to terminate and time out (instead of waiting
indefinitely) and continue with exit processing. This wait should address the original issue raised
by the prior commit and so it is okay to revert these two module changes from that commit.