[#205] Generate usable core files for raised fatal signals (e.g. SIGSEGV/SIG-11) in multi-thread processes
Problem statement
-----------------
In a multi-thread process, if a SIG-11 happens in say the TP worker thread, the signal handler
generic_signal_handler() is invoked. That in turn notices the current thread is not the MAIN
worker thread and so forwards the SIG-11 from the TP worker thread to the MAIN worker thread.
And generic_signal_handler() is invoked again in the MAIN worker thread. It is this invocation
that will call gtm_fork_n_core() for the fatal SIGSEGV signal. And since gtm_fork_n_core() does
a fork and then dumps the core file, it is the MAIN worker thread's C-stack that will be captured
in the core file (a fork only inherits the current thread's C-stack) making the core file unusable
since we are interested in the SIG-11 that happened in the TP worker thread.

Fix
---
As part of the FORWARD_SIG_TO_MAIN_THREAD_IF_NEEDED macro invoked by the TP worker thread (in the
above example on a SIG-11), after it forwards the signal to the MAIN worker thread, it does not
immediately return but waits for a signal (through a new global variable "safe_to_fork_n_core")
from the MAIN worker thread to indicate it is safe to do a "gtm_fork_n_core" call from the TP
worker thread even though it is not the MAIN worker thread. The MAIN worker thread pauses execution
while this core dump happens in the TP worker thread and then continues with its cleanup.

A new macro MULTI_THREAD_AWARE_FORK_N_CORE is invoked by the MAIN worker thread in generic_signal_handler()
wherever it needs to do a gtm_fork_n_core(). This macro checks if this is a RAISED signal (e.g. SIGSEGV,
SIGILL etc.) and if so sets safe_to_fork_n_core to TRUE and waits for this to be reset to FALSE (will be
reset by the TP worker thread or whichever thread got the SIGSEGV and is in the
FORWARD_SIG_TO_MAIN_THREAD_IF_NEEDED macro). If it is not a RAISED signal, this macro does a gtm_fork_n_core()
in the MAIN worker thread itself.