Fix one more build warning from gcc LTO on AARCH64
When keeping LTO enabled but temporarily using ld (instead of gold), the build on 64-bit ARM showed
an error in save_gv_cur_region variable being multiply defined and true enough it had two STATICDEF
definitions.  So that is now fixed by renaming one of the two definitions (one in gvn2gds.c) with
a gvn2gds_ prefix to avoid the duplicate definition.