[#205] Restore non-YottaDB signal handlers only after YottaDB exit handlers have been invoked
The v62002/gtm6638 subtest failed once in a while on some systems with the following diff.

39c39,40
< Pass
---
> Alarm clock
> Fail: expected=559 actual=

This test runs simplethreadapi 3n+1 for a while. As part of recent changes for YottaDB signal
handlers to co-exist with non-YottaDB signal handlers, there is now code in ydb_exit() that
resets the signal handlers to their non-YottaDB versions. But this is done BEFORE invoking
gtm_exit_handler() in the MAIN worker thread in case ydb_exit() is invoked in some other thread in
a SimplethreadAPI process. This means that it is possible that a SIGALRM timer is still active
at the time we reset the SIGALRM signal handler in ydb_exit() but before gtm_exit_handler()
(which does a cancel_timer()) has been invoked. If due to timing scenarios, this timer actually
pops before the cancel_timer() is done, the non-YottaDB SIGALRM handler will kick in. And I think
the system default handler for SIGALRM prints the "Alarm clock" message and just exits the process
(without also invoking YottaDB exit handler). This is most likely what caused the test failure.

The flow in ydb_exit() has been reworked so the signal handler reset happens AFTER the MAIN
worker thread has exited (i.e. after it has invoked gtm_exit_handler()).

Also noticed there was a pre-existing race condition in ydb_exit() (with multiple concurrent
invocations from different threads) which is now addressed by ensuring we hold the ydb engine
thread lock for the entire duration of the ydb_exit() even in the SimpleThreadAPI case
(when the now-nixed "wait_for_main_worker_thread_to_die" variable was TRUE).

Additionally, noticed the "struct sigaction" structure is not memset() to 0 before setting
the SIGALRM handler in init_timers() in gt_timers.c. Fixed that too just in case it can cause
other issues.