[#362] Avoid data differences between primary and secondary in case of kill -9 on primary side
The first commit of #362 was done in 2018 (SHA 438aabe7ba634d45ade6c6bdf3513c23ae60b742) where
the manually_start/dual_fail2_no_ipcrm2 subtest failed with a data difference. In that commit,
mur_back_process.c was fixed so it treats an automatically generated NULL record as a broken
transaction only in case of a fetchresync rollback.

But we had another failure of the same manually_start/dual_fail2_no_ipcrm2 subtest yesterday
where the automatically generated NULL record did not get treated as a broken transaction
because this NULL record was generated on side B where B was the secondary when the kill -9
happened (with no ipcrm) and the NULL record was generated by a non-fetchresync rollback
because B was coming up as a primary (as part of a failover). Side A which was a primary
before B was crashed, had a valid non-NULL record for the same seqno and since the rollback
on Side B did not treat the NULL record as a broken transaction, it played that forward on B.
This meant side B had a NULL record whereas side A had a non-NULL record corresponding to the
same seqno resulting in data discrepancy between the two sides once B started as the primary
replicating to A.

In thinking about the original commit of #362, it is not clear why the change was restricted
to only a fetchresync rollback. It seems more correct (and safer) to treat it as a broken
transaction in all cases. That will address the most recent test failure too.