Reduce gtmsecshr executable size (real exe, not wrapper) (#387 changes were missed out on this executable)
While at this, remove commented out executables libgtmcrypt and libmumps.  Also remove libgtmtls
and maskpass since they are part of the encryption plugin which is not distributed as part of the
YottaDB release binary tarball. They need to be built from source anyways at the customer environment.

The size of $ydb_dist/gtmsecshrdir/gtmsecshr on x86_64 reduced from 1Mb to 0.5Mb due to this change.