[#418] ydb_file_id_free()/ydb_file_is_identical()/ydb_file_name_to_id() and ydb_file_id_free_t()/ydb_file_is_identical_t()/ydb_file_name_to_id_t() issue PARAMINVALID error if input filename/fileid pointer is NULL
The code previously used to assume that the fileid input parameter was non-NULL so if a NULL pointer was
passed, one would get a SIG-11.

In case of ydb_file_name_to_id(), a NULL filename parameter caused a return immediately with no
fileid determination.

In both cases, the code now issues a PARAMINVALID error. This is safer/better.