[#420] Change all SimpleThreadAPI functions to get ydb engine lock and invoke corresponding SimpleAPI function directly (instead of going through the MAIN worker thread); To significantly speed up SimpleThreadAPI
Currently SimpleThreadAPI (e.g. a ydb_get_st() call) involves back and forth communication
with a MAIN worker thread through multiple system calls. That is a lot of overhead. The goal
of #420 is to avoid these system calls by doing the threaded API call in the current thread
itself. For this, the current thread needs to get the YottaDB engine lock to ensure it is the
only thread running the engine at one point in time. This commit drives us towards that goal.