[#423]  Eliminate a slight chance for a stack corruption when an interrupt occurs during execution.
Because ARM64 has 8 registers for passing arguments to a function, if more
than 8 arguments need to be passed, the stack is used. But in all these
functions, that stack is modified before calling another function. A new
value must be passed and room must be made in the parameter registers so one
parameter register must be moved to the stack. This entails copying the
existing stack, adding a new entry, and shifting the parameter registers.
ARM64 requires that pushes and pops to the stack go in 16 byte chunks (2 64
bit values at a time). The original attempt used a general purpose register
and wrote on 8 byte value at a time, starting at the current stack pointer.
When all the 8 byte values were on the stack, the stack pointer was set to
point at the proper location. Unfortunately, during the stack copy time, the
stack pointer wasn't always pointing to the real top-of-stack, only being
updated when the copy was complete. If an interrupt occurred during this
interval when the stack pointer wasn't correct, the interrupt handling
routines could using the stack at the stack pointer corrupting values.

To fix the problem, the stack pointer is used to push the values - making sure
that everything is done 16 bytes at a time.

Other cosmetic changes included in this commit are

* Remove the line of code that includes the file "stack.si". There are no usages of
  stack.si macros in the current version of these files. The macros existed in prior
  versions, but were eliminated as the code evolved. Removing the references to stack.si
  should have been done earlier, but wasn't.

* Remove these files as they are no longer used.

  ci_restart.s - should not have been included in the first place. It was
  mistakenly created because an i386 assembly file of the same
  existed.

  stack.si - due to code revisions and simplifications, the macros in this
  file were no longer used anywhere in the assembly source code.
  There is no reason to keep it.

* Change the first argument passed to gvcmz_neterr() to be a 64 bit zero.
  Currently, the argument, w0, is only 32 bits as is the zero register, wzr.

* Change the branch instruction to b.ne for consistency with the other conditional
  branch instructions in the code.

* Change the reference to the 64 bit register X0 to x0 (upper case X to lower
  case x) for consistency with the other 64 bit registers in the assembly code.