[DEBUG_ONLY] Enhance read_regions() in gtmsource_readfiles.c in debug builds to capture more information in case a given seqno will never be found (TR_WILL_NOT_BE_FOUND)
We occasionally see test failures where the source server fails an assert in debug builds because
it decided a given seqno can never be found in any of the journal files even though that seqno
is later seen to be present in one of the journal files. The assert that fails is the following.

%YDB-F-ASSERT, Assert failed in sr_unix/gtmsource_readfiles.c line 1922 for expression (!*brkn_trans || (ydb_white_box_test_case_enabled && ((WBTEST_REPLBRKNTRANS == ydb_white_box_test_case_number) || (WBTEST_MURUNDOWN_KILLCMT06 == ydb_white_box_test_case_number) || (WBTEST_JNL_FILE_LOST_DSKADDR == ydb_white_box_test_case_number))))

From the core file, we do not have enough information to analyze the issue since the logic that
led to this conclusion involves multiple regions and journal files. Therefore, the code is enhanced
for debug builds to capture more trace information as we iterate through each region in the
read_regions() function trying to locate a given seqno. This will help us better analyze such
assert failures in the future.