[#420] Implement generic signal handling forwarding; Invoke deferred signal handler in MAIN worker thread or in YottaDB engine lock holding thread just before releasing thread lock
continue_handler()/jobexam_process()/jobexam_signal_handler rely on info and context being filled in
but in SimpleThreadAPI environment they will not be due to signal forwarding to main worker thread.
It is a pre-existing issue in SimpleThreadAPI in r1.24 (#205) that is now fixed. Not considered
likely to be encountered in practice so no user-visible issue created for this separately.