[#420] Fix deferred_signal_handler() so it runs fine even if we are not in the MAIN worker thread in SimpleThreadAPI mode
It is okay to do so because deferred_signal_handler() is invoked from YottaDB runtime code and so
the caller already owns the multi-thread YottaDB engine mutex lock. This function is never invoked
through a signal handler and so we do not have the usual issue with signal handlers being invoked
in a different thread than the one currently running the YottaDB engine.