[#420] Remove assert that can now fail since in SimpleThreadAPI mode it is possible that STAPI_SET_SIGNAL_HANDLER_DEFERRED gets called in a signal handler while we are in start_first_timer()
It seems safe to remove the assert since even if the assertion fails, it means we have a case
where timeroot is NULL (i.e. no timers are scheduled) but deferred_signal_handling_needed is non-zero
which means we will clear this variable a little later. Small extra work but no correctness issue.