[#420] Avoid duplicate FORCEDHALT messages in .mje files due to multiple invocations of generic_signal_handler() for the same signal due to signal forwarding in SimpleThreadAPI
This fixes a test failure in the v60000/gtm4525b subtest where the symptom was that one or
more *.mje* files contained TWO lines with FORCEDHALT messages instead of the expected ONE.
This is possible if the first invocation noticed the need to defer exit processing and printed
the FORCEDHALT message and the second invocation was because of a forwarded signal (e.g. from
the MAIN worker thread).

With the fixes, the second invocation will skip logging the FORCEDHALT message as well as
bumping exit_state (i.e. skip all operations which are not safe to redo).