Handle pre-existing rare edge case in incorrectly clearing jnl file io writer pid
We had an assert failure in the dual_fail_extend/dual_fail2_mustop_sigquit subtest in one run.

> %YDB-F-ASSERT, Assert failed in sr_port/jnl_write_attempt.c line 149 for expression ((&jb->io_in_prog_latch)->u.parts.latch_pid == writer)

This turns out to be a longstanding edge case issue. Once we find the CURRENT_JNL_IO_WRITER(jb)
is non-zero and matches our local copy of "writer", we grab crit if necessary and then
check if the process id "writer" is no longer alive. If so we go ahead and clear the
jb->io_in_prog lock. But it is possible the process cleaned up the jb->io_in_prog lock
and terminated normally just before we got crit. In that case, the jb->io_in_prog lock
could be currently owned by some other process and we would end up clearing that
incorrectly paving the way for this lock to be concurrently held by more than one
process at the same time.

This is considered a very rare edge case that no user-visible issue is created for this.