[#420] Fix SLEEP_USEC macro calls that occur in sleep loops with a max timeout to sleep for exactly the time requested instead of returning prematurely
This is an issue that existed even before but #420 exposed it by causing the MAIN worker thread to
wake up the YottaDB engine multi-thread lock holding thread every 1 microsecond until a signal
handler was driven.

The reason these SLEEP_USEC macro calls need to be fixed is that once the sleep loop times out
then it could invoke recovery procedures that assume a fixed time of sleep has elapsed. This is no
longer guaranteed if the SLEEP_USEC macro call had FALSE as the second parameter.