[#420] Fix hiber_start_wait_any() so it does not restart sleep if an EINTR occurs
A recent #420 commit (SHA eca650321492316df7e7f79920bac1a19cc19037) caused the regression
by changing the SHORT_SLEEP macro to always wait for specified time even in case of an EINTR).
hiber_start_wait_any() is used by op_lock2() etc. all of which rely on the function returning
as soon as any interrupt occurs. Not doing this could cause op_lock2() etc. to take a lot
longer than usual particularly if the same operations are done a lot of times (like is done
in the relink/barrage subtest which ran a lot longer and failed due to this regression).

Additionally, all calls of SHORT_SLEEP were now examined to see if they have the potential of
issues due to the recent macro changes. Only hiber_start_wait_any() was identified as needing
change and is now fixed in this commit.