[#420] While waiting for instance to be unfrozen, invoke deferred signal handler only AFTER checking exit_state and invoking EXIT if appropriate
The v60000/gtm4525c subtest hung without this fix because a mumps process was inducing the
DBBMLCORRUPT error which caused the instance to be frozen. This mumps process later hung while
trying to halt due to the instance freeze. It was killed by the test using a MUPIP STOP but
that in turn invoked the deferred signal handler which in turn caused the exit handler to be
driven and that caused the instance to be re-frozen (due to the white-box test case
WBTEST_ANTIFREEZE_DBBMLCORRUPT being enabled for that portion of the test). The test had just
before done the instance unfreeze so it was expecting a REPLINSTUNFROZEN message in the source
server log which never showed up (and the test hung) because of the re-freeze.

This commit avoids the re-freeze issue by invoking the deferred signal handler after the
exit_state check. In case of a MUPIP STOP (SIGTERM) or any other process-terminating signal,
the EXIT() invocation does an _exit() which bypasses all exit handlers if we are already in
an exit handler and that avoids invoking exit handlers multiple times unnecessarily (which
as a side effect also avoids the duplicate freeze that causes the v60000/gtm4525c subtest hang).