[#420] Avoid possible deadlock in ydb_init() if invoked in SimpleThreadAPI mode while inside TP
ydb_init() currently returns right away if "ydb_init_complete" is TRUE AND if there is no need
to create a nested callin frame. But if there is a need to create a nested callin frame, it would
fall through to code which gets a multi-thread YottaDB engine lock ydb_engine_threadsafe_mutex[0]
(corresponding to tptoken YDB_NOTTP). This means that if such a ydb_init() call happens while
inside of TP (e.g. inside the callback function), it would be deadlocked. This is because, in TP,
the lock ydb_engine_threadsafe_mutex[0] would already be held by the thread that did the ydb_tp_st()
call and potentially locks corresponding to ydb_engine_threadsafe_mutex_holder[1] and higher indices
could also be held by one or more threads depending on the current depth of the TP ("dollar_tlevel"
global variable). And so waiting for the lock ydb_engine_threadsafe_mutex_holder[0] in ydb_init()
while inside TP would cause a deadlock in this thread in case the TP callback function at this TP
depth is waiting for this ydb_init() call to return (i.e. ydb_engine_threadsafe_mutex_holder[0]
will not be available until the TP is done).

To avoid this deadlock, ydb_init() is now fixed to return right away if "ydb_init_complete" is TRUE.

But this presents two issues.

1) Code that tries to create a nested callin frame using call-ins needs to invoke ydb_nested_callin()
   and needs to issue a CIMAXLEVELS error in case the callin depth exceeds the max allowed (10
   currently).  This error is issued in ydb_nested_callin() (which is currently called from ydb_cij()
   and ydb_ci_exec()) but at that point we do not have any condition handler established. This
   means the CIMAXLEVELS error value is never returned from ydb_cij() or ydb_ci_exec(). To fix this,
   a new macro named INVOKE_YDB_NESTED_CALLIN_AND_RETURN_ON_ERROR is introduced in sr_unix/gtmci.c.
   This establishes a condition handler (pre-existing gtmci_ch()) and uses that to detect an error
   and if so returns that error from ydb_cij() and ydb_ci_exec().

2) Code that tries to create a nested callin frame while using SimpleAPI/SimpleThreadAPI needs to
   invoke ydb_nested_callin(). This was previously accomplished
   a) for SimpleThreadAPI by the LIBYOTTADB_RUNTIME_CHECK or LIBYOTTADB_RUNTIME_CHECK_NORETVAL macros
      which are invoked as part of every SimpleThreadAPI call.
   b) for SimpleAPI by the LIBYOTTADB_INIT macro which did a LIBYOTTADB_RUNTIME_CHECK macro invocation
      which is invoked as part of every SimpleAPI call.
   The LIBYOTTADB_RUNTIME_CHECK/LIBYOTTADB_RUNTIME_CHECK_NORETVAL macros invoked ydb_init()
   even if the current frame is not of type SFT_CI. But now that ydb_init() does not invoke
   ydb_nested_callin(), that check is unnecessary and so is removed. Instead, the LIBYOTTADB_INIT
   and LIBYOTTADB_INIT_NORETVAL macros, which are invoked by every SimpleAPI function (and by
   every SimpleThreadAPI function after it has already got the needed YottaDB engine multi-thread
   lock), are now enhanced to check if the current M frame is not of type SFT_CI and if so invoke
   ydb_nested_callin() (just like the INVOKE_YDB_NESTED_CALLIN_AND_RETURN_ON_ERROR macro does).

Additionally, it was noticed that ydb_ci()/ydb_cip()/gtm_ci()/gtm_cip() were skipping the
SIMPLEAPINEST check portion of the LIBYOTTADB_INIT macro but otherwise duplicating what the macro did.
Therefore the LIBYOTTADB_INIT macro has been moved to a new LIBYOTTADB_INIT2 macro which allows for
an additional parameter to be passed indicating whether the SIMPLEAPINEST error check has to be performed
or not. All existing usages of LIBYOTTADB_INIT are unaffected because this macro is redefined to be
an invocation of LIBYOTTADB_INIT2 with the 3rd parameter set to DO_SIMPLEAPINEST_CHECK_TRUE. And all
LIBYOTTADB_RUNTIME_CHECK usages in ydb_ci()/ydb_cip()/gtm_ci()/gtm_cip() are now replaced with the
new LIBYOTTADB_INIT_NOSIMPLEAPINEST_CHECK macro which now invokes LIBYOTTADB_INIT2 with the 3rd parameter
set to DO_SIMPLEAPINEST_CHECK_FALSE.