[#420] Avoid possible deadlock in ydb_exit() if invoked in SimpleThreadAPI mode while inside TP
The main changes are in ydb_exit.c to issue INVYDBEXIT error if we are in TP and in SimpleThreadAPI
mode.

While doing that, noticed scope for adding a few asserts since ydb_exit() returns a positive error
code (this is how it was inherited from GT.M and hence we could not change this to negative error
codes like is the case for SimpleAPI/SimpleThreadAPI functions). Made similar changes to ydb_init()
as well since that is similar in that regard to ydb_exit().