[#370] Implement new SimpleAPI utility functions ydb_ci_tab_open()/ydb_ci_tab_open_t()/ydb_ci_tab_switch()/ydb_ci_tab_switch_t()
The main change is to now maintain a linked list of open call-in tables.  A new typedef
"ci_tab_entry_t" describes a call-in table. The new global variable TREF(ci_table_all) points to
a linked list of these structures each corresponding to a call-in table.

The prior global variables TREF(ci_table) and TREF(callin_hashtab) were pointing to two aspects of the
same call-in table and those are now removed and instead a new global variable TREF(ci_table_curr)
points to the "ci_tab_entry_t" structure corresponding to the current call-in table. Within
that structure, the "cname_list" member gives access to the "callin_entry_list" linked list that
TREF(ci_table) previously pointed to.  And the "hashtab" member gives access to the hash table that
TREF(ci_hashtab) previously pointed to. Similarly, TREF(ci_filter_table) and TREF(ci_filter_hashtab)
are now replaced by one global variable TREF(ci_table_internal_filter).

A new function ci_tab_entry_open() is defined in sr_unix/exttab_parse.c to open a call-in table.
That is now invoked by various portions of ydb_cij() and ydb_ci_exec() which eliminates a lot of
duplicate code.  This maintains the TREF(ci_table_all) global variable.

The pre-existing function citab_parse() now accepts an additional parameter "fname" which if non-NULL
points to the file name that defines the call-in table (instead of getting the file name from the
env var "ydb_ci").

The new function ydb_ci_tab_open() invokes ci_tab_entry_open() with a non-NULL "fname" to open the
desired call-in table. The new function ydb_ci_tab_open_t() is a SimpleThreadAPI wrapper on top of
the ydb_ci_tab_open() function.

The new function ydb_ci_tab_switch() switches the current call-in table (TREF(ci_table_curr)) based
on the input table handle. The new function ydb_ci_tab_switch_t() is a SimpleThreadAPI wrapper on
top of the ydb_ci_tab_switch() function.

The new global variable TREF(ci_table_default) points to the call-in table, if any, that was opened
through the "ydb_ci" environment variable.

While at this, a longstanding issue with ydb_cip()/gtm_cip() has also been fixed.  And that is,
they never filled in the descriptor like was documented. This is because they invoked ydb_ci_exec()
passing in ci_info->handle as the second parameter. This parameter was passed by value so there
was no way ci_info->handle was updated to a non-NULL value after the first call to ydb_ci_exec()
when the descriptor was determined. This meant ydb_cip() was not any faster than ydb_ci() all this
while. This is now fixed by passing in "ci_info" as the parameter so the handle can be filled in.