[DEBUG_ONLY] Fix incorrect assert introduced in #350 (failed after integrating GT.M V6.3-006)
After integrating GT.M V6.3-006 into the YottaDB master branch, in a terminal,
key in "mupip upgrade" and hit <Enter>. It failed with the following assert.

%YDB-F-ASSERT, Assert failed in sr_unix/iott_readfl.c line 1238 for expression (!1 || IS_SETTERM_DONE(io_ptr))

The failure C-stack was the following.

 #0  __pthread_kill () at ../sysdeps/unix/sysv/linux/pthread_kill.c:56
 #1  0x00007f5c9ef1cc03 in gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2  0x00007f5c9ef1e12b in gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3  0x00007f5c9eefad61 in ch_cond_core () at sr_unix/ch_cond_core.c:80
 #4  0x00007f5c9efef55c in rts_error_va () at sr_unix/rts_error.c:192
 #5  0x00007f5c9efeea6a in rts_error_csa () at sr_unix/rts_error.c:99
 #6  0x00007f5c9f4a994a in iott_readfl () at sr_unix/iott_readfl.c:1238
 #7  0x00007f5c9f4a149d in iott_read () at sr_unix/iott_read.c:26
 #8  0x00007f5c9f27b397 in op_read () at sr_port/op_read.c:70
 #9  0x00007f5c9ef02f57 in cli_get_parm () at sr_unix/cli_parse.c:1025
 #10 0x00007f5c9eefb8cd in cli_get_str () at sr_unix/cli.c:271
 #11 0x00007f5c9f1fa1a1 in mupip_upgrade () at sr_port/mupip_upgrade.c:115
 #12 0x00007f5c9eebf406 in mupip_main () at sr_unix/mupip_main.c:124
 #13 0x000055ba15601913 in dlopen_libyottadb () at sr_unix/dlopen_libyottadb.c:148
 #14 0x000055ba15601240 in main () at sr_unix/mupip.c:19

This is because cli_get_parm() now does an op_read() call which means even MUPIP can go through
iott_readfl() which means the RESETTERM_IF_NEEDED macro invocation done at the end of iott_readfl()
will assert fail because it was not written to be in sync with the SETTERM_IF_NEEDED macro.

The SETTERM_IF_NEEDED macro skips the setterm() call if IS_GTM_IMAGE is FALSE.
But the RESETTERM_IF_NEEDED macro has an assert that does not take this into account.
This is now fixed.