[#434] ydb_exit() while inside an external call no longer terminates process with a SIG-11 fatal error
gtmci_ch() is a condition handler that is safe to be used only when the code surrounded by the
condition handler creates a call-in frame (either the first call-in or a nested call-in frame).
In ydb_exit(), neither of these conditions are true. And so, if a INVYDBEXIT error occurs,
the condition handler gtmci_ch() will incorrectly unwind the M stack that would later cause
a SIG-11 when the ydb_exit() call returns to the M caller of the external call.

This is fixed by instead using ydb_simpleapi_ch() which is almost the same as gtmci_ch(),
but has no such M-stack unwinding issues.