[V63006] Fix fallout from GTM-4263 code changes (Entering more than one .m file name at "What file:" prompt should work)
The r120/ydbdist subtest invokes "$ydb_dist/mumps" without any arguments.
And when prompted for a file to compile, it hits <Enter>. And that assert failed as follows.

> $ydb_dist/mumps
What file:
%GTM-F-ASSERT, Assert failed in sr_unix/zl_cmd_qlf.c line 98 for expression (*srclen)

This happened after the V63006 code changes integration. This is fixed in zl_cmd_qlf.c by
checking for *srclen and if it is 0, to return instead of proceeding. The caller (gtm_compile())
issues a FILENOTFND error in that case.

While investigating this change, I realized that if one keys in a list of file names in the
"What file:" prompt, it does not work.

> $ydb_dist/mumps
What file: x.m y.m
%GTM-E-NOTMNAME, x.m y is not a valid M name
%GTM-E-ZLNOOBJECT, No object module was produced

The file name is treated as "x.m y.m" instead of "x.m" and "y.m". This is fixed by enhancing
cli_get_parm() (in cli_parse.c) to see if a list of values is specified and if so return only
the first value and store the remaining in a list of command-line values to be processed
(cli_lex_in_ptr->tp).