[V63006] Fix more fallout from GTM-4263 code changes (mupip trigger -select when prompted for "Output file:" displays the file name TWICE)
With GT.M V6.3-006, in a terminal run the following without specifying an output file for the trigger select.
You will be prompted for it. Once you enter the file name there, you will see the file name displayed again
in a new line. And the shell prompt showing up right after the second file name (in case there are no
triggers in the database).

> mupip trigger -select
Output File: x.xx
x.xx>

The above is considered ugly output by me. Instead, the below is what I would expect to see.
The file name x.xx is printed only once and the shell prompt appears in the beginning of a new line.

> mupip trigger -select
Output File: x.xx
>

The duplicate file name printing is fixed by changing the RESETTERM_IF_NEEDED and SETTERM_IF_NEEDED
macros to also work for MUPIP now that iott_readfl() etc. can be invoked for MUPIP in cli_get_parm()
using the op_read() call. This takes care of disabling the echo mode in the terminal while user
input is being read in and avoids duplicate display of user input.

The extraneous newline is fixed in sr_unix/trigger_select.c by removing an extraneous PRINTF of a newline
which would go through the stdout device which is no longer correct to use since V6.3-006 since
we need to do M IO and not C buffered IO on stdout. Note that a newline is already printed after the
input "x.xx" is keyed in in cli_get_parm() using an op_wteol() in a prior commit so that is good enough.